+++ includes/Specials/SpecialEditLexicon.php
@@ -195,7 +195,7 @@ class SpecialEditLexicon extends SpecialPage {
 		foreach ( $fields as $field ) {
 			$name = $field['name'];
 			$value = $request->getVal( $name );
-			if ( $value !== null ) {
+			if ( $value !== null && $value !== '' ) {
 				// There's no extra conversion logic so default values
 				// are set to strings and handled down the
 				// line. E.g. boolean values are true for "false" or
@@ -315,7 +315,12 @@ class SpecialEditLexicon extends SpecialPage {
 					'type' => 'hidden',
 					'name' => 'word',
 					'default' => $word
-				]
+				],
+				'consumerUrl' => [
+					'name' => 'consumerUrl',
+					'type' => 'hidden',
+					'default' => $this->getRequest()->getText( 'consumerUrl' )
+				],
 			],
 			'formId' => 'syncForm',
 			'submitMessage' => $this->msg( 'wikispeech-lexicon-sync-error-button' ),
@@ -350,7 +355,12 @@ class SpecialEditLexicon extends SpecialPage {
 			'page' => [
 				'name' => 'page',
 				'type' => 'hidden'
-			]
+			],
+			'consumerUrl' => [
+				'name' => 'consumerUrl',
+				'type' => 'hidden',
+				'default' => $this->getRequest()->getText( 'consumerUrl' )
+			],
 		];
 		return $fields;
 	}
@@ -399,6 +409,13 @@ class SpecialEditLexicon extends SpecialPage {
 			'options' => $itemOptions,
 			'default' => ''
 		];
+
+		$fields['consumerUrl'] = [
+			'name' => 'consumerUrl',
+			'type' => 'hidden',
+			'default' => $this->getRequest()->getText( 'consumerUrl' )
+		];
+
 		return $fields;
 	}
 
@@ -577,11 +594,7 @@ class SpecialEditLexicon extends SpecialPage {
 		$this->modifiedItem = $item;
 
 		if ( array_key_exists( 'page', $data ) && $data['page'] ) {
-			// @todo Introduce $consumerUrl to request parameters and
-			// @todo pass it down here. Currently we're passing null,
-			// @todo meaning it only support flushing local wiki
-			// @todo utterances.
-			$this->purgeOriginPageUtterances( $data['page'], null );
+			$this->purgeOriginPageUtterances( $data['page'], $data['consumerUrl'] ?? null );
 		}
 
 		return true;