+++ includes/Utterance/UtteranceStore.php
@@ -543,6 +543,13 @@ class UtteranceStore {
 		int $pageId
 	): int {
 		$remoteWikiHash = self::evaluateRemoteWikiHash( $consumerUrl );
+
+		if ( $consumerUrl ) {
+			$this->logger->info( __METHOD__ . ": Flushing utterances for page $pageId at $consumerUrl" );
+		} else {
+			$this->logger->info( __METHOD__ . ": Flushing utterances for page $pageId" );
+		}
+
 		$dbw = MediaWikiServices::getInstance()->getConnectionProvider()->getPrimaryDatabase();
 		$results = $dbw->select( self::UTTERANCE_TABLE,
 			[ 'wsu_utterance_id' ],